requirements.txt
environs=9.4.*
django==4.0.*
python_telegram_bot==13.10